finagle-mux: Interrupt writes in server dispatcher
Problem

When writing a response in mux, we lose track of the writer
promise (i.e. write and forget). This is unfortunate because
we want to attempt to interrupt the write if we receive a
cancellation from the client. For example, this becomes
increasingly important with large responses that are more likely
to timeout.

Solution

Compose writes as part of our pending state on the mux server.
This has some implications which are subtle. Namely, because
netty3 doesn't guarantee that write closures will be executed on
the event loop thread, it means that we can now race with a mux
client which reuses tags eagerly. To mitigate this, we change
the way we do writes in ChannelTransport which closes the window
for one potential race. Additionally, we make the server more
lenient by servicing duplicate tags and forcing pathalogical
clients to deal with it.

RB_ID=802514