finagle-core: Do not trace when Tracer is null
Problem

Both `StatsReceiver` and `Tracer` may be set to null in Finagle server. It
makese sense to skip all the tracing- and stats-related filters in that
case while materializing the stack.

Solution

Introduce `Tracer.isNull` so we can check that in every filter that does
tracing and skip it if needed.

RB_ID=809574