finagle-http2: Cleartext http2 transport
Problem

We want to support http2

Solution

The netty http2 codec is still experimental, so for now, in the
name of getting this out the door, we're going to use their http
=> http2 converter, which will have a slightly smaller featureset
than if we were using the http2 api directly.  We can't hook it up
to finagle quite yet because there isn't a conversion from finagle
Request / Response to netty4 Request / Response, but when it lands,
we can start hooking it up on the client side.

One example of the slight weirdness of the netty http2 codec is
that it's incompatible with turning off autoread.  You have to call
a method on the connection handler to get it working.

Other than not doing backpressure, this doesn't handle waiting for
settings properly, so in theory there's a race condition with
upgrading, but since this is experimental, and I want to get this
out the door so we can start experimenting with listeners, I am OK
with shipping it as is.

Result

We can play with an http2-themed finagle Transport!

RB_ID=807769