finagle-core: Added Round Robin load balancer and improved Simulation.scala
Problem

We wanted to compare Round Robin load balancing with the choices
available in Finagle.

Solution

We added a simple round robin loadbalancer to the finagle-core package
along with a unit test suite.

We also added improvements to Simulation.scala such as printing
individual latency numbers, adding a slow client in the middle of a
simulation run.

We added a .roundRobin() function to Balancers.scala as well as a
newline to .aperture()'s comment to improve the scaladoc output.

We added a type constraint to DistributorT#This to enforce by the
compiler what was being done by convention.

We changed protection modifiers to several member classes and traits of
loadbalancer package to allow running their tests from sbt.

Result

  * Added Round Robin load balancer for testing purposes.
  * Modified Simulation to allow a slow client in the middle.
  * Modified Simulation to allow tracing latency per request.
  * Added a shell script to automatically run all the Simulations I care about
  * Added some R code in a literate file to help me build nice graphs.
  * Added a newline for improved scaladoc in Balancers.scala

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

RB_ID=814447