Problem
Mux shares a session with multiple tag streams but offers no mechanism
to ensure that an individual stream is not dominating a session.
This is particularly problematic in the presence of large payloads
which can cause head-of-line blocking for an entire session. Ideally,
we would share a session uniformly across streams to increase
goodput.

Solution

In order for a tag stream to yield, we need to defer flushing its
writes to the OS buffer. This implies that we add some some latency
to writes. In netty3, a call to write will immediately flush the
bytes to the OS buffer if there is space (which is the common case).
For example, the following will sit in a tight loop and flush all
the fragments to the OS buffer:

def write(fragments: Iterator[ChannelBuffer]): Future[Unit] =
 if (!fragments.hasNext) Future.Done else
 trans.write(fragments.next()).before(write(fragments))

In other words, trans.write is usually synchronous. To address this
we queue and flush writes periodically (note, netty4 introduces
distinct operations for writing (enqueueing) and flushing). The
current implementation flushes writes on read and uses a scheduled
timer to ensure that we flush regularly if throughput is low. As
mentioned earlier, this does entail a latency penalty. Because we
are now buffering payloads in mux, we get nice cancellation semantics.
This does require that we introduce a new mux message type (Rdispatch)
to assure the sender of a Tdispatch they can safely clear their
buffers.

Any fragment larger than `com.twitter.finagle.mux.maxFrameSize` is
fragmented into `maxFrameSize` sized chunks. The window size is
fixed in the current implementation and future work may include a
flow control algorithm which dynamically sizes it.

Results

The resuls are extracted from a simple client/server setup where
the server responds with 500kb payloads for each request (i.e. all
of our responses are fragmented). The request rate is such that we
are not saturating our NIC allocation (< 37.5 MB/s).

Ping p99: With fragmentation, we see smoother ping latencies and
up to a 50% decrease during peaks.

CPU: With fragmentation, we see ~20-30% less system CPU usage on
both clients and servers (likely due to less syscalls since we batch
writes).

Request p99: Request latency does take a big hit and increases by
~80%.  Future work includes experimenting to see if we can do better
by batching writes on the event-loop thread and avoid having to set
a separate cadence on the `flow-writer` thread.

Allocations/request: no statistical difference between the two

RB_ID=794641