finagle-redis: Changes to support PubSub do not compile in 2.10.6
Problem

The current code has several issues which do not compile in Scala
2.10.6. In the Transactions trait, the value "svc" is shadowed and
failing compilation in 2.10.6 (recursive call without a defined
return type). Additionally, in Client the passed in factory is not
visible for subclasses which mix in the trait.

Solution

Disambiguate the proxied "svc" in the Transactions trait and add
typing. Make the factory constructor arg in Client visible to
subclasses.

Result

Code that compiles in both Scala 2.10.6 and 2.11.7.

RB_ID=815777