Allow wildcard elements in Dtab prefixes
Problem

In some cases, it is convenient to simply ignore parts of a path when
processing dtabs. This is currently quite awkward to accomplish with
custom namers.

Consider the case when a name like the following is processed:

   /http/1.1/GET/host/resource

Solution

Alter the Dtab syntax to support wildcards in prefix paths. This enables
dtabs in the form:

   /http/1.1/* => /h

This would allow the above name to easily be processed to
/h/host/resource.

Signed-off-by: Antoine Tollenaere <atollenaere@twitter.com>

RB_ID=813914