finagle: Add `andThen` overloads for Filter.TypeAgnostic
Problem
-------

When composing Filter chains, it's often necessary to add boilerplate
type annotations that make the code harder to read, even when there is
only one possible type.

Solution
--------

Add method overrides for composing `Filter.TypeAgnostic` with `Filter`,
`Service`, etc. This enables seamless mixing of `Filter` and
`Filter.TypeAgnostic`, which will be specialized to the required type
when it is composed.

RB_ID=811012