finagle-stats: Ensures that there are histograms immediately when we call counts
Problem

No histograms show up when you first query the /admin/histograms
endpoint.  This is because we cache the endpoint, and only refresh
minutely when we read /admin/metrics.json, the same way metrics
does.

Solution

Eagerly compute the histogram the first time you hit the endpoint
so we have something to see.

RB_ID=819459