Verify that histogram snapshotting works after a lapse in snapshotting
Problem

When snapshotting does not occur for serveral minutes, the 60-second
snapshotting is not honored. If snapshotting does not occur for two minutes,
histogram windowin will not be properly honored until snapshot has been called
twice.

Solution

Change snapshot windowing to be based on the current time.

Fixes twitter/twitter-server#33.