finagle-mux: refactor writes in MuxFramer using Offer/Broker
Problem

Our current implementation for flushing writes in MuxFramer
doesn't scale well with the number of connections in a process.
The flush timer spins unneccessarily. More importantly, any flush
cadence we choose that is out of band with non-fragmented writes
will make it difficult to ensure that we are acheiving our primary
goal of stream equity over the session.

Solution

Reimplement our write logic in terms of Offer/Broker. We have three
processes that need to synchronize:

1. Add a new stream (i.e. an incoming write)
2. Interrupt an active stream.
3. Flush active streams in a round-robin fashion.

This is naturally modeled using offer/broker and allows us to remove
any mention of the execution mechanics. When we have a new stream to write,
we begin the `writeLoop` and concurrent streams can join an
active `writeLoop`.

RB_ID=819520