finagle-netty: Fix resource leak in http proxy connect handler
Problem

As discussed on Github [1], it's not only a good idea to interrupt a handshake
(either SSL/TLS handshake or HTTP proxy handshake) on interrupted connect future
(eg: interrupted by `TimeoutFactory`) it's absolutely necessary to do in order to
prevent resource leaks from channels remaining open.

Imagine the following scenario:

1. A client sends an HTTP proxy handshake message `CONNECT /twitter.com`
2. A server accepts that but never sends a response back
3. A client stalls the connection waiting for a response from a server
4. `TimeoutFactory` kicks in and interrupts the connect promise
5. Finagle stack will create a new channel for a next request, but this will remain open

Solution

Close the channel and fail pending writes when a connect promise is interrupted.

Note that this RB also fixes the behavior when channel is closed by remote peer
so we no longer need to wait for a handshake to complete.

[1]: https://github.com/twitter/finagle/issues/345#issuecomment-210754044

RB_ID=823927