finagle-netty4: Introduce SslConnectHandler
Problem

We need to have a proper tooling around delaying the transport connection until
the SSL/TLS handshake is done and the session is verified.

Solution

`SslConectHandler` reuses the ideas introduced in `HttpProxyConnectHandler` and delays
the propagation of the connect promise back to a transport until SSL handshake is
succeed.

RB_ID=824682