finagle-core: Document components of TraceId
Problem

The distinction between trace id and span id is not obvious to
people reading the API documentation; the contrast between the trace
id and span id components of a TraceId is limited to the word
"particular".

Solution

Provide further explanation of the components of TraceId with
examples.

RB_ID=825871