finagle-toggle: Introduce new feature toggle module
Problem

As library owners, we don't sufficient control over how slowly a
change is rolled out into production.

Solution

Introduce the concept of feature toggles into Finagle. This commit is
the first part of more to follow and as such it is not terribly useful
yet.

Result

While the APIs are experimental and subject to change, this is a good
building block for the future work to be done.

RB_ID=822194