Lazily instantiate StackServer's JvmFilter
Problem

StackServer statically instantiates a `MkJvmFilter` with a `c.t.jvm.Jvm`. This
creates a timer thread that samples JVM stats. We'd like to disable the
JvmFilter stack module during profiling, but this is not currently possible
since stats polling is initiated regardless of whether the module is used.

Solution

Lazily instantiate StackServer's JvmFilter.

Result

The stats polling timer is only initiated when the JvmFilter server module is
enabled.

Also-by: Steve Jenson <stevej@buoyant.io>