Requeue retryable failures
Problem

Response classification has an explicit mechanism for marking failures as
retryable, yet this response classification is not taken into account when
determining whether a request should be requeued.

Solution

Provide RequeueFilter with a ResponseClassifier.  RequeueFilter considers a
request to be requeued when its response is classified as a RetryableFailure.