finagle-netty4: Do not automatically release buffers
Problem

Buffers are being stored on a queue to be handled by different threads, but
immediately afterwards they were freed. When other thread tried to work with
them, an Exception occurred.

Solution

Do not release the buffers, as they will be accessed in the future.
This is done by changing a flag to disable automatic release.

Signed-off-by: Daniel Schobel <dschobel@twitter.com>

RB_ID=830197