finagle-core: remove the use of keytool in TLS setup
Problem:
Finagle shells out to keytool as part of its loading of keys for TLS setup.

The Finagle TLS setup code currently calls keytool to convert a pkcs12 file to a java keystore for subsequent loading and use in TLS. This is redundant, as directly loading pkcs12 files as keystores has been supported in the JVM for some time. Also, it assumes that all processes being launched have JAVA_HOME in their path and pointing to a JDK (instead of a JRE).

Solution:

The call to keytool has been removed and the code now directly loads the pkcs12 file. Other minor code changes to use more appropriate classes for handling this data and updated comments to refelct these changes.

RB_ID=832070