finagle-http: Removes duplicates for StreamTransport versions
Problem

We had a bunch of older versions of different Http building blocks
that weren't on StreamTransport yet, because the codecs weren't cut
over.

Solution

Cut over the codecs, and replace the old building blocks with the
new ones.  We still need to find good homes for a few new ones in
exp, but at least we no longer have duplicates anymore.

RB_ID=830150