finagle-core: MonitorFilter installs parameterized monitor
Problem

Previously, MonitorFilter would fail a request if a synchronous
exception was thrown anywhere within the service.  This led to
surprising results, since a callback that seemingly didn't have an
effect on the normal execution, like logging a failure, could fail
the entire request if they threw an exception.

Solution

Change MonitorFilter to instead use the parameterized Monitor, which
is NullMonitor by default.  This matches the behavior described in
the scaladoc, and has a less surprising result.

Result

Finagle services are more deterministic, and code that is added
just for side-effects will not spuriously fail requests.

RB_ID=832979