finagle-core: BufWriter implementors only implement `arrayToWrite`
Problem

As we add different writers and writing methods to BufWriter, it would be nice
if we didn't have to implement all those methods in subclasses.

Solution

All writing methods in BufWriter now use `arrayToWrite(bytes: Int)`, which is
implemented by the subclasses and returns the array to write to and the starting
index.

RB_ID=833394