finagle-core: Add a FailureAccrualPolicy that marks an endpoint dead when getting only failures for a specified amount of time.
Problem

We don't have a FailureAccrualPolicy that allows us to mark an endpoint dead when getting only failures for a specified amount of time. FailureAccrualPolicy.consecutiveFailures(Int, Stream[Duration]) does something similar, but it depends on the service QPS.

Solution

Add a FailureAccrualPolicy that marks an endpoint dead when the success rate in a specified time window is under the required threshold.

RB_ID=829984