finagle-core: Prevent unbounded memory leak from FailFastFactory
Problem

`FailFastFactory` has a singleton `Stream` that is used as the default
for backoffs. Due to the way `Stream` works, if there is a
"permanently" dead host we leak a `Stream.Cons` every time we issue a
backoff.

Solution

Cap the backoffs after a reasonably sized number (16) then use a
constant backoff which prevents the space leak.

(discovery of the bug, along with the fix, is courtesy of vkostyukov).

RB_ID=834328