finagle-thrift,thrift-mux: ResponseClassification of unencoded failures
Problem

Response classification for Thrift and ThriftMux did not handle the
case where a failure is not Thrift-encoded. For example, both
`BackupRequestLost` and `IndividualRequestTimeoutException` are
produced on the client side and thus will not be encoded as a
`Return`.

Solution

Ask the classifiers to handle these without deserialization.

Result

Response classification works as users expect it should.

RB_ID=834212