finagle-kestrel: Do not empty cluster when it becomes Pending
Problem
--

Some service discovery implementations (most notably Zk2Resolver) use the Pending state to signal
that the underlying mechanism is currently unavailable (e.g. ZooKeeper is not reachable). Kestrel
MultiReader interprets it as an empty set of Addresses, causing the cluster to empty when service
discovery is not reachable.

Solution
--

Do not empty the server set when Addr goes to Pending.

RB_ID=832923