finagle-http: Add HTTP Status Code 451
Problem

RFC7725, "An HTTP Status Code to Report Legal Obstacles" was published
in February 2016. It is a valid HTTP/1.1 response code, but is not
currently a part of com.twitter.finagle.http.Status.

Solution

Add it to com.twitter.finagle.http.Status (for Scala) and
com.twitter.finagle.http.Statuses (for Java).

RB_ID=834997