finagle-{memcached, kestrel}: Use BufWriter.dynamic for encoding
Problem

For a Netty 4 enabled memcached client/server, the message encoder
in the pipeline will need to be in terms of Buf.

Solution

Re-write the encoder in terms of Buf and add a Buf => ChannelBuf
handler after it in the pipleline. We plan to unify the conversion across
protocols but this is a first step.

RB_ID=829643