finagle-toggle: Add constraints on ids
Problem

Toggles lack constraints on what their ids can be. Without constraints
there is a high likelyhood of emoji ids. While this may sound
entertaining, it will almost certainly not be for future mainainers.

Solution

Add restrictions on what is allowed to be `A-Za-z0-9._-` and update
existing unit tests to conform.

RB_ID=836828