finagle-core: Unracify PipeliningDispatcher
Problem

Some memcached users were reporting out-of-order responses in a
pipeline, but in a tcp dump, the responses all seemed correct.  This
suggested there was a bug in the client, and in particular around
the order in which responses are given.

Although it's unlikely, it's possible that writes could race, so
that one write finishes before the other, but enqueues its Promise
later.

Solution

Serialize the write and enqueue step, so that it's no longer possible.

Result

Pipelined protocols are deracinated.

RB_ID=834927