Move HTTP Context filters into the Http client and server stacks
Problem

The headers used to propagate context information are hardcoded into the Http
client and server, provided by `Http.{Client,Server}.newDispatcher`.

In linkerd, we would like to admit linkerd-specific headers, and so we would
prefer to replace these filters.

Solution

Insert a ServerContextFilter module at the top of the server stack and a
ClientContextFilter module at the bottom of the client stack. Include the
DtabFilter in the ServerContextFilter module since it extracts similar
contextual information.