finagle-core: Introduce StatsTransport
Problem

As we move more logic into transports, we need to have a principled way of recording
useful metrics at that layer. Today, we have `ChannelStatsHandler` to collection metrics
at the netty layer, `StatsFilter` for the service layer, and nothing for the layer that
glues the two together.

Solution

Create StatsTransport with an initial implementation which records failures in read/write.

RB_ID=839664