finagle: Removed all of the guava caches from finagle
Problem

Guava is a very widely used dependency, so it's a pain to OSS users of finagle to have to deal with the dependency in finagle.

Solution

As a first step, let's remove all of the uses of guava cache.  We can replace them with Caffeine, which is a spiritual successor and much simpler.

Result

Many fewer dependencies on Guava!

RB_ID=833970
TBR=true