finagle-core: Move HTTP Context filters into the Http client and server stacks
Problem

The headers used to propagate context information are hardcoded into the Http
client and server, provided by Http.{Client,Server}.newDispatcher.

It can be useful to to admit other specific headers, so it should be possible to
replace those filters.

Solution

Insert a ServerContextFilter module at the top of the server stack and a
ClientContextFilter module at the bottom of the client stack. Include the
DtabFilter in the ServerContextFilter module since it extracts similar
contextual information.

Signed-off-by: Antoine Tollenaere <atollenaere@twitter.com>

RB_ID=840600