finagle-http2: Hook up http-specific netty stuff
Problem

We have a bunch of http-specific netty stuff that we hadn't hooked
up to http2 yet.

Solution

Hook it up.  One interesting impedance mismatch was that since we
pass pipeline manipulations as a ChannelInitializer, it's easiest
to reason about the rest of the initializations as a ChannelInitializations
too.

RB_ID=841987