finagle-core: Interrupt pipelined failures after waiting for a long time
Problem

We never cut connections on timeouts, even if they're hung.

Solution

Wait after a timeout to see if the pipeline ever clears–if it
doesn't, but continues to back up, cut the connection.

Result

Even if a pipelined connection hangs on a single request, we're
able to restart the connections and begin to make progress again.

RB_ID=827731