finagle-toggle: Add optional comment field in JSON
Problem

It is useful to be able to leave comments in your configuration files
to explain why decisions were made. Unfortunately, JSON doesn't
support inline comments.

Solution

Add an optional "comments" field to the "toggle" hash which is ignored
in the creation of the `ToggleMap` and serves a similar purpose.

RB_ID=843410