Ensure that Transport.collate's future is satisfied before its reader is closed
Problem

There is a subtle race condition between DelayedReleaseService and
HttpTransport. DelayedReleaseService may be notified that a stream has
completed before the underlying transport has observed this update. This allows
services to be released to the CachingPool before HttpTransport has an
opportunity to update its ConnectionManager's state. This may manifest as a
ConnectionClosedException, when a closed transport is returned as valid by
CachingPool.

Solution

Update Transport.collate to satisfy its closure promise before actually closing
the response stream. This allows the ConnectionManager to observe the state
update before a response's Readers are notified.

Fixes twitter/finagle#517