finagle-core: Remove Call of setEnableSessionCreation on SSLEngines
Problem

For SSLEngines created by and passed into Finagle for use with TLS,
the setEnableSessionCreation method is called on the engine with a
value of true. true is the default for JSSE implementations, and
finagle-native treats this call as a no-op. However, for other native
engines this call can be an unsupported operation, because there is no
available way to turn off session creation.

Solution

Remove this call from the internals of finagle, so that exceptions
are not thrown when trying to call setEnableSessionCreation on
SSLEngines which do not support it.

RB_ID=845765