finagle-toggle: Introduce environment into StandardToggleMap
Problem

A safer rollout for a toggle in many cases is to enable the feature
for staging environments before turning it on in production.

Solution

Introduce `ServerInfo.environment` which can report the enviroment a
server is running in. This is utilized by `StandardToggleMap` to
prefer environment-specific JSON configuration before
non-environment-specific JSON configuration.

Result

Library owners will be able to enable Toggles in a safer manner.

RB_ID=845877