finagle-memcached: Bugfix memcached Zipkin integration
Problem

When getting a value from memcached the cache hit/miss annotation
was submitted after the cr annotation was submitted.

Submitting the cache hit/miss annotation after the cr annotation resulted
in the span being recreated and ending up in the DeadlineSpanMap.
When the deadline exceeded the span was submitted a 2nd time overwriting
the correct duration of the original span.  Once that happened
the span was also annotated with 'finagle.flush'.

Solution

Update the memcached specific tracing filter so memcached specific
annotations are submitted in between cs and cr annotations.

Result

The memcached span is not re-submitted a 2nd time and does not end
up in DeadlineSpanMap. The duration is correct when viewed in zipkin-web
and there is no 'finagle.flush' annotation anymore.

Signed-off-by: Jillian Crossley <jillianc@twitter.com>

RB_ID=847035