finagle-toggle: ToggleMap.orElse should keep Metadata descriptions
Problem

Most `Toggle.Metadata` does not have a `description` as the
`ToggleMap` implementations don't have this data. When we combine
`ToggleMaps` with `orElse` we should try to keep `description` when we
can.

Solution

Continue to prefer the rest of the `Metadata` from `self` but use
descriptions from `that` if its not defined.

RB_ID=847122