Add transactions
This introduces a method `Client#inTransaction`, which accepts a function from `Client` to `Future[T]`.  The operations on the passed `Client` will be executed on a new `Client` with a single connection (created with `ServiceFactory.const`).  If the resulting future is successful, the transaction will be committed; if it fails, the transaction will be rolled back.  The temporary single-connection client is then closed.