finagle-toggle: Make JSON description optional
Problem

For a service owner's JSON configuration files, it is detrimental to
require the description field. This is because the library owner has
defined the toggle in their configuration file and as such, requiring
the service owner to also add a description is wasteful and could lead
to confusion in the future if they input something incorrect.

Solution

Ignore the description field for service owner's config files.

Also, in preparation for future work, `ToggleMap.newMutable` and
`ToggleMap.Proxy` have been made public.

RB_ID=848340