twitter-server: Admin endpoint for Toggle mutations
Problem

Service owners need an API for updating the in-memory Toggles.

Solution

Provide PUT/DELETE methods on /admin/toggles so that
create/update/deletes can be performed.

Result

To create or update the in-memory version of a toggle:
`PUT /admin/toggles/$libraryName/$id?fraction=$fraction`

To remove the in-memory version of a toggle:
`DELETE /admin/toggles/$libraryName/$id`

RB_ID=849906