Close service when prepared query can't be parsed
This fixes an issue wherein the service isn't closed when the parse
phase of a prepared statement fails.  This causes that connection to
hang.  Instead, the parse failure should cause the service to be closed
(just as the successful completion of the query closes the service, and
the failure at any other point during the process does the same).