finagle-http: Close server transport when request is not keepalive.
Problem

When a finagle-http server receives a non-keepalive request (i.e. with
Connection: close), it does not close the underlying transport. Instead it
waits for the client to teardown the connection. This causes undesirable
behavior: client connections go into TIME_WAIT on the remote side since they do
not receive a FIN from the serverside. This can cause clients to exhaust ports
when not using keepalive connections.

Solution

The HttpTransport should be closed eagerly, after it is finished processing a
request/response that is not keepalive (i.e. it is HTTP/1.0 or the `Connection`
header is set to `close`).