finagle-http2: Handle goaways properly client-side
Problem

Clients swallow GOAWAYs instead of following normal client behavior.

Solution

As a temporary measure, we're going to satisfy all oustanding
requests with a fabricated 4XX based on the GOAWAY we received.

We also changed the http2 end to end tests to test the http2 case,
which will be the common case, instead of the http2 upgrade, which
is interesting but not as urgent as making sure http2 works.

Result

Http2 clients don't hang randomly anymore.

RB_ID=852038