finagle-exp: Abstract out parts of the DarkTrafficFilter for potential re-use
Problem

Parts of the DarkTrafficFilter are useful to be defined in one place for re-use.
Additionally, the current stat scoping is inconsistent with current Finagle best
practices so has been changed from camelCase to snake_case ("darkTrafficFilter" ->
"dark_traffic_filter").

Solution

Create an AbstractDarkTrafficFilter trait which encapsulated potentially
re-usable functionality.

Result

If another filter with a different type signature is created it can re-use the
important components of the DarkTrafficFilter through sub-classing the
AbstractDarkTrafficFilter.

RB_ID=852548