finagle-core: Refactor loadbalancer's Distributor code
Problem

The user guide lacked information on the load balancer's behavior when
all nodes are down. While working on documenting this, I found the
`DistributorT` implementations to have a lot of repetition.

Solution

Refactor much of the shared code up into `DistributorT` and then document
the behavior in the user guide that Finagle uses a node it believes to be
down.

RB_ID=852565