finagle-zipkin-core: Remove package protection for RawZipkinTracer
Problem

RawZipkinTracer is a base class that is extended to support other transports
like http or kafka. The current implementation requires subtypes to be in the
com.twitter package. This isn't likely intentional.

Solution

This removes the constraint that requires such subtypes to be in the
com.twitter package.

Result

RawZipkinTracer can be extended from other packages, such as zipkin.finagle

Signed-off-by: Jordan Parker <jparker@twitter.com>

RB_ID=853430