finagle-core: Remove Java 6 compatibility code from FinagleScheduler
Problem

FinagleScheduler.switchToBridged uses the class loader to attempt to instance a
j.u.c.LinkedTransferQueue, and on failure, doesn't continue with switching the
scheduler. This is a compatibility layer for Java 6. We no longer support
Java < 8, so this check is no longer necessary.

Solution

Create the class instance using the new operator.

RB_ID=855000