finagle: Remove while loop in P2C load balancer pick function
Problem

The P2C load balancer trait uses a while loop to attempt to ensure that
it doesn't randomly pick the same node twice for the following comparison logic
, and gives up after 10 retries. It is possible for this to fail (although
unlikely, ~1 in 20 million and the repercussions are not very serious).

Solution

Replace the loop with logic that is guaranteed to pick a different node in a
single step.

RB_ID=855635