finagle-core: Reduce number of useless rebuilds for P2C/Aperture
Problem

Finagle's randomized LBs (i.e., Aperture, P2C/LeastLoaded, P2C/EWMA) are implemented
in a way that they partition the input replica set to distinct `down` and `up` nodes
so that only `up` is used by the distributor.

While this approach is pretty straightforward, it comes with an additional cost in
re-partitioning (allocations/GC, CPU) when a) one of the down nodes is up b) one
of the up nodes is down. The picture is getting worse when a service with large
cluster doing a rolling restart of its nodes/replicas with big batches.

For example, with batch size 10% of the 100-nodes cluster, the chances of hitting a
dead node with P2C are 19%. We either:

 (a) pick two deads (10/100 * 9/99) or
 (b) pick one dead and one alive (10/100 * 90/99) or
 (c) pick alive and then dead (90/100 * 10/99)

Technically, this means that every 5th request will be triggering LB rebuilds
when 10% of the cluster is down. Good news - LBs have some logic to collapse this
kind of updates, and in fact, this helps a lot. But given the high rate of
update-requests, the huge chunk of them is still reaching the LB, causing extremally
high resources usage.

Solution

Let's use math/probability to see how often LBs will be hitting dead nodes if
they **don't partition at all**.

Let's have a look at the extreme case when 50% of the cluster is down. With P2C
and max-effort = 5 (number of times we're willing to retry on LB), the
probability of returning a dead node out of half-dead cluster after 5 attempts
is 0.1% (0.25^5, where 0.25 is the probability of picking two dead nodes in row).

This means that if the half of the cluster is down, P2C will be able to mitigate
that and only return sub-optimal choices (when a better choice may have been
available) for about 0.1% (1/10 of the percent) of requests. That doesn't mean
that 0.1% of requests will be failed by a given client, because there is an
additional layer of requeues involved above those "bad picks".

Now let's have a look at the less extreme case, 10% of the cluster is down. With
max-effort = 5, the probability of hitting a dead node is 0.00000000001% (almost
zero).

With Aperture, those probabilities are even less since they consist of two
dependent actions - 1) including a dead node into an aperture and 2) picking
a dead node out of the aperture. In practice, we've seen that Aperture mitigates
situations when 10% of the cluster is dead much better than P2C. Math also
confirms that.

**Putting it all together**, instead of relying on partitioning (with the cost of
CPU, GC, and rebuild-storms) in mitigating temporary unavailability of individual
nodes, this RB suggests to put math (not computers) into work and using the fact
that the probability of hitting a temporary dead node is extremely low,
**do not rebuild the server set when the dead node is seen** in P2C/Aperture.

This not only reduces the number of LB rebuilds (allocations, CPU) when cluster
is doing a rolling restart, but also helps in a general case (there is always a
dead node somewhere in the cluster that we will never hit, but it trigger
useless rebuilds).

One way or another, LBs are still relying on the underlying namer to trigger
rebuilds and exclude permanently unavailable nodes (restarted, unannounced, etc)
from the server set.

Result

Our experiments showed that services using this change mitigated rolling restarts
of their dependencies (CPU and GC were flat) and also improved overall p99 latency
by up to ~20%.

RB_ID=855030