finagle: IllegalArgumentException in Aperture sort
Problem

The new Aperture rebuild function attempts to sort the nodes by Status, but
this can change during the sort, resulting in an
IllegalArgumentException from the underlying java TimSort algorithm.

Solution

Only perform a true sort on the token, which is immutable, and then
perform a 'counting sort' based on the mutable status, discarding any
Status transitions that occur during the sort.

RB_ID=856861