finagle-http2: Client-side cleartext downgrades
Problem

We want to be able to turn on http2 on the client-side before the
server-side has upgraded.  This will allow us to upgrade without
checking first if servers have upgraded, so upgrading doesn't have
to be phased.  This also unblocks potentially turning on http2
automatically for users, ie with toggles.

Solution

We capture upgrades, surface them to the transporter layer, and use
that to inform how it creates connections.

One subtle part is what to do when an upgrade is in progress.
Instead of buffering them, I've chosen to initiate them as http/1.1
connections, so that they can make progress even if the upgrade
takes a while (for example if the request that initiates the upgrade
is really wide).

This has a few results.  First, it means that clients can only talk
to servers that can downgrade to http/1.1 if it doesn't want to
buffer its requests.  This is required by the spec, but in practice
some servers can't.

We may change this in the future, especially when ALPN comes on
line, and moves protocol negotiation into the connection setup.

Result

Clients will upgrade or downgrade automatically and don't need input
from users.

RB_ID=856637