finagle-serversets: disable serverset2 Addr stabilization if windows are 0 seconds
Problem:

There isn't an efficient way to disable stabilization in `Zk2Resolver`.
You can set the `batchWindow` and `removalWindow` to 0 seconds but
this causes the epoch to trigger on every tick which seems wasteful.

Solution:

Check the corner case where `batchWindow` and `removalWindow` are
both 0 seconds. Return the `Addr` as is.

RB_ID=859276