Problem
When a finagle-http server receives a non-keepalive request (i.e. with
Connection: close), it does not close the underlying transport. Instead it waits
for the client to teardown the connection. This causes undesirable behavior:
client connections go into TIME_WAIT on the remote side since they do not
receive a FIN from the serverside. This can cause clients to exhaust ports when
not using keepalive connections.

Solution

HttpServerDispatcher should terminate connections eagerly when a client requests
that the connection be closed.

Signed-off-by: Jordan Parker <jparker@twitter.com>

RB_ID=854542