finagle-serversets: consolidate serverset path parsing
Problem:

Serverset paths are parsed in an ad-hoc manner in each namer (e.g.
each of them manually splits a string by ":").  This can be cleaned
up by introducing a layer of abstraction (ServersetPath) and sharing
logic.

Result:

There is now a single place where serverset paths are parsed, which
makes future changes to path syntax easier.  This patch will parse
shard ID but it will be acted on in a future patch.

RB_ID=862740