finagle: Add module to Nack AC filter and replace multiplier param with success rate threshold
Problem

The `NackAdmissionFilter` companion object does not have a module, so
service owners cannot add it to their clients' stacks. Also, the multiplier
param is harder to understand and configure than it needs to be.

Solution

Add a module to the companion object, and replace the `multiplier` param with a
`successRateThreshold` param. This param is the inverse of the multiplier and it
represents the lowest success rate the filter tolerates before probabilistically
dropping requests. E.g., with a `successRateThreshold` of `9D/10`, the filter
will probabilistically drop requests when the success rate falls below 90%.

Results

Now service owners can add the filter to their client stacks and can configure
the filter more easily.

RB_ID=861931